+++ example_subscribe.py
@@ -93,12 +93,18 @@ binance_websocket_api_manager.subscribe_to_stream(stream_id, markets=markets_4)
 time.sleep(1)
 binance_websocket_api_manager.get_stream_subscriptions(stream_id)
 
-time.sleep(10)
+time.sleep(2)
 channels_2 = ['trade', 'kline_5m', 'kline_15m', 'kline_30m', 'kline_1h', 'kline_12h', 'depth5']
 binance_websocket_api_manager.unsubscribe_from_stream(stream_id, channels=channels_2)
-time.sleep(3)
-binance_websocket_api_manager.get_stream_subscriptions(stream_id)
+request_id = binance_websocket_api_manager.get_stream_subscriptions(stream_id)
 
+while binance_websocket_api_manager.get_result_by_request_id(request_id) is False:
+    print("Wait to receive the result!")
+    time.sleep(0.5)
+
+print(str(binance_websocket_api_manager.get_result_by_request_id(request_id)))
+
+time.sleep(10)
 while True:
     #binance_websocket_api_manager.print_summary()
     binance_websocket_api_manager.print_stream_info(stream_id)